+++ webf/change log.md
@@ -2,4 +2,4 @@
 
 1. 安装包 npm install --save file:../packages/network_manage
 2. 增加路由 network-manage
-3. 在新路由中增加3 个按钮，分别是查询 allNetworks, evmNetworks and 查询 id=3 的 network
+3. 在新路由中增加3 个按钮(纵向居中)，分别是查询 allNetworks, evmNetworks and 查询 id=3 的 network